GCP/himas-federated-agents/pyproject.toml
[project]
name = "himas-federated-agents"
version = "0.2.0"
description = ""
authors = [
    {name = "Yash Khare", email = "khare.y@northeastern.edu"}
]
readme = "README.md"
requires-python = ">=3.12,<3.14"
dependencies = [
    "google-adk>=1.13.0,<2.0.0",
    "google-adk[a2a]",
    "google-genai>=1.32.0,<2.0.0",
    "google-cloud>=0.34.0,<0.35.0",
    "google-cloud-bigquery>=3.11.0,<4.0.0",
    "google-cloud-aiplatform>=1.111.0,<2.0.0",
    "google-cloud-logging>=3.5.0,<4.0.0",
    "google-cloud-storage>=2.19.0",
    "tensorflow>=2.15.0",
    "protobuf>=5.28.3,<7.0.0",
    "scikit-learn>=1.3.0,<2.0.0",
    "db-dtypes>=1.2.0,<2.0.0",
    "pandas>=2.0.0",
    "cryptography>=41.0.0,<42.0.0",
    "python-dotenv>=1.0.0,<2.0.0",
    "requests>=2.31.0,<3.0.0",
]

[project.optional-dependencies]
dev = [
    "google-cloud-aiplatform[adk,agent-engines,evaluation]>=1.93.0",
    "pytest>=8.3.5",
    "pytest-asyncio>=0.26.0",
    "google-adk[eval]>=1.5.0",
]
deployment = [
    "absl-py>=2.2.1",
]

# The build system is updated to use hatchling, a standard build backend.
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["agent_hospital_a"]
